% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demon.R
\name{demon}
\alias{demon}
\title{The example for GSparO}
\usage{
demon()
}
\description{
demon is a function that implements GSparO for an example of least squares regression with A and b being Gaussian ensembles. A figure plotting the true signal and estimation by GSparO is illustrated in Plots, and the errors of least squares regression and obtained solution are printed. Two packages ThreeWay and ggplot2 should be installed for implementing demon.
}
\details{
Copyright by Dr. Yaohua Hu, College of Mathematics and Statistics, Shenzhen University.
Email: mayhhu@szu.edu.cn
}
\examples{
demon()
}
